Fix tew check style warnings
