+++ plugin/src/test/java/io/jenkins/plugins/analysis/core/model/DetailsTableModelTest.java
@@ -55,29 +55,8 @@ class DetailsTableModelTest extends AbstractDetailsModelTest {
             builder.setDescription("See <a href=\"https://clang.llvm.org/extra/clang-tidy/checks/bugprone/forward-declaration-namespace.html\">Clang-Tidy documentation</a>.");
             var issue = builder.build();
             
-            var descriptionProvider = new DescriptionProvider() {
-                @Override
-                public String getDescription(final Issue issue) {
-                    return issue.getDescription();
-                }
-
-                @Override
-                public String getCategoryUrl(final Issue issue) {
-                    // Extract URL from description
-                    var description = issue.getDescription();
-                    if (org.apache.commons.lang3.StringUtils.isNotBlank(description)) {
-                        var hrefPattern = java.util.regex.Pattern.compile("href=[\"']([^\"']+)[\"']");
-                        var matcher = hrefPattern.matcher(description);
-                        if (matcher.find()) {
-                            return matcher.group(1);
-                        }
-                    }
-                    return "";
-                }
-            };
-            
-            var model = new TableRow(createAgeBuilder(), createFileNameRenderer(), descriptionProvider, issue,
-                    createJenkinsFacade());
+            var model = new TableRow(createAgeBuilder(), createFileNameRenderer(), new TestDescriptionProvider(), 
+                    issue, createJenkinsFacade());
             
             assertThat(model.formatPropertyWithUrl("category", issue.getCategory(), issue))
                     .isEqualTo("<a href=\"https://clang.llvm.org/extra/clang-tidy/checks/bugprone/forward-declaration-namespace.html\" target=\"_blank\" rel=\"noopener noreferrer\" title=\"View documentation\">bugprone-forward-declaration-namespace</a>");
